+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * [BUGFIX] Renamed label `state` to `name` on `node_systemd_service_restart_total`. #1393
 * [BUGFIX] Fix netdev nil reference on Darwin #1414
 * [BUGFIX] Strip path.rootfs from mountpoint labels #1421
+* [FEATURE] Add new thermal_zone collector #1425
 
 ## 0.18.1 / 2019-06-04